refactor powermodels :sheep:
